# api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "這是「花中查詢」的API文件檔，您可以使用這個API來進行對於成績系統的查詢。",
    "version": "1.0.0",
    "title": "花中查詢API文檔",
    "contact": {
      "email": "yurisakadev@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "{server}/api",
      "variables": {
        "server": {
          "default": "https://hlhsinfo.ml"
        }
      }
    }
  ],
...
  "paths": {
...
    "/getScheduleList": {
      "get": {
        "tags": [
          "資料查詢"
        ],
        "summary": "查詢當前使用者所有課表",
        "description": "",
        "operationId": "scheduleList",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/scheduleList"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "dataAuth": []
          }
        ]
      }
    },
    "/getSchedule": {
      "get": {
        "tags": [
          "資料查詢"
        ],
        "summary": "查詢當前使用者的課表",
        "description": "",
        "operationId": "schedule",
        "parameters": [
          {
            "in": "query",
            "name": "class",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "使用者的班級(可以由`/getScheduleList`取得)"
          },
          {
            "in": "query",
            "name": "teacher",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "使用者的導師(可以由`/getScheduleList`取得)"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/schedule"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "dataAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "login": {
        "type": "http",
        "scheme": "bearer"
      },
      "dataAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "scheduleList": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "schedules": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "class": {
                      "type": "string"
                    },
                    "teacher": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "url": {
            "type": "string"
          }
        },
        "xml": {
          "name": "scheduleList"
        }
      },
      "schedule": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "object",
              "properties": {
                "section": {
                  "type": "string"
                },
                "class": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "className": {
                        "type": "string"
                      },
                      "location": {
                        "type": "string",
                        "nullable": true
                      },
                      "teacher": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                },
                "time": {
                  "type": "object",
                  "properties": {
                    "start": {
                      "type": "string"
                    },
                    "end": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "url": {
            "type": "string"
          }
        },
        "xml": {
          "name": "schedule"
        }
      },
...
    }
  },
...
}